requirements.txt
Django
psycopg2
Pillow
djangorestframework
djangorestframework-simplejwt
drf-spectacular

